requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
apscheduler==3.11.0
playwright==1.52.0
httpx
pydantic==2.11.5
google-genai==1.19.0
python-dotenv==1.1.0
browser-use
langchain-deepseek
